backend/requirements.txt
Django==4.2.5
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
django_simple_plausible==0.0.4
gunicorn==21.2.0
ipython==8.15.0
